+++ src/components/elements/my-segment/index.tsx
@@ -1,0 +1,2 @@
+export { default } from './my-segment'
+export type { MySegmentOption, MySegmentProps, MySegmentSize } from './type'